Reorganize parallel module using classes
